requirements.txt
alembic
argon2-cffi
bcrypt
click
celery
celery-redbeat
redis
amqp
limits
first
pyramid
pyramid-services
pyramid-retry
pyramid-tm>=1.1.1,<2.0.0
gunicorn
pyramid-jinja2
wtforms
sqlalchemy
psycopg2
zope.interface
zope.sqlalchemy
msgpack-python
passlib
